feat(cli): install pinned nightly toolchain automatically (#1541)
`cargo openvm build` will use `rustup` to check if the required nightly
toolchain and `rust-src` component are installed, and will install them
if they are not. If this step fails, then the entire build process
errors (I wasn't sure if this was the preferred behavior or if we should
let the rest of the cargo build attempt continue, but it felt like for
earlier stopping and debugging it's better to fail the process).

Updated only the cli workflow because in multithreaded tests, I don't
want it to try to install the toolchain in multiple threads
simultaneously.

closes INT-2911