fix: Update Mod Builder symbolic expression multiplication test (#1551)
The symbolic expression multiplication unit test was being run with an
incorrect expected result. The test became out-of-date after #1473
changed the `max_abs` method.

This PR updates the test with the correct expected result and adds the
Mod Builder crate's unit tests to the CI workflow.

This change has no impact on the functionality of the Mod Builder crate
as it only updates the tests.