feat: make rv32im testing uniform accross all the chips (#1630)
Made the rv32im tests pass and made all the testing files to have the
same testing interface.
Deleted the `test_adapter`. Kept all the test cases unchanged. The only
commented test case remaining is the `store` test to the address space
4, which is failing because currently memory accesses with block size 4
are not supported with the address space 4.

All the test files have 3 types of tests: Positive, Negative, and Sanity
tests.
All the test files have 2 helper functions: `create_test_chip`,
`set_and_execute`.

An important thing to notice about negative tests when expecting an
interaction fail (aka ChallangePhase error) is that ther might be an
imbalance created for the wrong reasons. For example, there might be an
imbalance on the range checker bus created by the interactions:
[send 1] (sent from the chip_air)
[receive 2] (the execution did `add_count(2)` at some point)
This is not a "valid" fail since 1 is still in the range of the range
checker. Because of this a manual check is needed for all the negative
checks. To see all the imbalances occurred during a test remove the
'disable_debug_builder();' line from the `run_negative_test` function
and run the test. I am 95% sure that I wen through all the negative
tests and checked that the imbalances occurred are correct.

The `test_adapter` tried to address this issue by getting rid of
interaction imbalances on the memory bus. But even with the `test
_adapter` a manual check was necessary.
To solve this I suggest that we somehow keep all the interactions that
occur during the test and automatically check that actually an invalid
interaction has happened on a specified bus.

Resolves INT-3975

---------

Co-authored-by: Ayush Shukla <ayush@axiom.xyz>