fix: Fix setting initial memory (#1690)
We sometimes call `set_initial_memory` even with `Volatile` interface
chip. Don't know if this alone is a good design, but inside this
function we reset the tracing memory to use 8 as `initial_block_size`
and then checked that in the volatile case the memory is empty, but the
initial block size was already overwritten. This change fixes it.