chore: add execution insn/s logging (#1764)
We currently do not run e1 in the local (single machine) proving flow.
We will update openvm-prof once that is added.

Due to the limited nature of the metric aggregation by segment,
`total_proof_time_ms` includes the e2(metered) time for single segments
(all aggregation proofs) but excludes it for app proofs. I include the
app e2 time in the final totals.

Updates the metrics caching and PR comment so that it will compare using
the PR's target branch as base.

Example run:
https://github.com/openvm-org/openvm/blob/benchmark-results/benchmarks-pr/1764/regex-f073953995b8735d92bf3a91b452e3047cb1d671.md