feat: `InterpretedInstance` holds pre-computed handlers (#1924)
For ultra-low latency, all data that can be pre-computed once the
program binary is known should be held in `InterpretedInstance` and not
re-computed on the fly. This PR does this and closes INT-4470

- `VmExecutor` is constructed once the VM config is known and holds all
data that can be determined before given a program binary. (For
execution this is only the executors)
- `InterpretedInstance` should be created once the program binary is
known and holds all pre-computed buffers and function pointers, so that
`execute` can focus on only runtime specific computations.
- Refactored so that the `InterpretedInstance` now provides the only
`execute` and `execute_metered` functions, and `VmExecutor` only
provides constructors for the `InterpretedInstance`, which is separated
between pure (E1) and metered (E2) via the `Ctx` generic.
- Also added the ability to `execute_from_state` for future E1 purposes.
- (Out of scope here) Note that E2 always executes until completion --
the current API seems not designed for any kind of async execution of
segments. I think to do this perhaps the easiest is if a separate thread
is created with view access to the `SegmentationCtx` which can poll for
new `Segment` creation and start off other processes.
- (Out of scope) I want to do a similar refactor for `VirtualMachine` vs
a new `VmInstance` type, but going to hold off and try to switch E3 to
use `InterpretedInstance` first

The codspeed benchmarks should be largely ignored since any perf
improvements are mostly from decreasing the pre-computation time.