fix(perf-metrics): perf instrumentation slow (#1936)
The per-instruction call to calculate memory access adapter heights is
very slow because it iterates linearly through the entire adapter
record. I changed it now to only do this at the end. The downside is
that we won't be able to see the contribution to access adapters on a
per-instruction basis, but this could be extracted more efficiently in a
different way.

Proof it works:
https://github.com/axiom-crypto/openvm-reth-benchmark/actions/runs/16767319820