feat: Parse struct name in complex field + ECC init macros as string (#1945)
Cherry-picked from https://github.com/openvm-org/openvm/pull/1891
without the twisted Edwards macros.

The `complex_init!` and `sw_init!` macros are updated so that instead of
parsing a `Path`, it parses a `LitStr`. This makes it so that the
`_init!` macros do not require any path imports in the file where it is
called, which greatly improves the developer experience.

The main files changed aside from test files and auto-generates
`openvm_init.rs` files are:
- extensions/algebra/complex-macros/src/lib.rs
- extensions/ecc/sw-macros/src/lib.rs

In both macros, the change essentially boils down to changing:
(before)
```rust
        let items = input.parse_terminated(<Expr as Parse>::parse, Token![,])?;
```
to (after)
```rust
        let items = input.parse_terminated(<LitStr as Parse>::parse, Token![,])?;
```
and then associated type changes (mostly simplifications) in converting
tokens to strings.

There is one additional change needed only for the `sw_init!` macro
case. Previously the linked function
```rust
extern "C" fn #setup_extern_func()
```
defined by the `sw_init!` macro did use the path of the struct within
the function implementation. After this PR, we avoid this import by
moving that part of the logic into the `fn set_up_once()` function
defined for the struct itself. This changed the function signature of
the extern function to:
```rust
            extern "C" fn #setup_extern_func(uninit: *mut core::ffi::c_void, p1: *const u8, p2: *const u8)
```
[Here](https://www.diffchecker.com/0ncTMwe2/) is the diff of the code
that was moved from the `#setup_extern_func()` into the `set_up_once()`
function.

---------

Co-authored-by: Avaneesh Kulkarni <avaneeshk@intrinsictech.xyz>