chore: return memory tests back (#1950)
This resolves INT-4550.

Merkle tests checked that, when we touch some memory,
- the finalized root that is obtained by "build a tree on initial
memory, then update with the touched memory" is the same as if we just
built the tree on the final memory,
- the merkle trace does the interactions we want it to.

I updated them so that they check the same but use the new architecture.

The general memory tests checked that we can do all kinds of accesses
(persistent/volatile, native/rv32 address space, etc etc) and nothing
breaks. Apparently, I did the same in `online.rs`, so I just moved my
tests to that file, where I think they indeed belong.